refactor(architecture): land MOD-SHARED-KERNEL at the module-graph bottom (pilot S5, ARCH-CHANGE-002)
Re-partitions the registry to break the WORKSPACE-IDENTITY <->
WORKTREE-LIFECYCLE value cycle recorded as ARCH-WAIVER-003:

- New MOD-SHARED-KERNEL owns src/models.ts, src/constants.ts,
  src/sessionAssignment.ts + src/worktreeSessionAssignment.ts (moved from
  workspaces/), and src/worktreeIdentity.ts (the WorktreeKey type and its
  canonical codecs, moved from worktrees/types.ts, which re-exports them
  so consumers keep their import paths). The kernel has no value-level
  outbound dependencies.
- Measured outcome: the workspace<->worktree pair dissolved and its waiver
  retires; the 13-module cyclic cluster shrank to 12 and no longer
  contains MOD-WORKTREE-LIFECYCLE; four more 2-cycles dissolved as return
  edges moved onto the kernel. Baseline regenerated (17 -> 12
  fingerprints); waiver ledger keeps its bijection.

mayDependOn refreshed from the measured graph. Pure moves and path
updates; no behavior change (codecs byte-identical, pinned by
worktreeKeyCodecs tests).
