+++ tests/unit/tooling/architectureGuards.test.js
@@ -55,7 +55,9 @@ function copyGuardFixture(t, mutationPath, mutate = source => source) {
         'src/models.ts',
         'src/todos/types.ts',
         'src/webview/dashboardViewModel.ts',
-        'src/workspaces/sessionAssignment.ts',
+        'src/sessionAssignment.ts',
+        'src/worktreeSessionAssignment.ts',
+        'src/worktreeIdentity.ts',
         'src/workspaces/types.ts',
         'src/worktrees/types.ts',
         'src/aiSessions/conversation/composition.ts',
